requirements.tf.txt
wheel
tqdm
matplotlib
tensorflow-gpu==2.3.0
tensorflow_datasets
tensorflow-text
